+++ jubatus/core/regression/normal_herd.hpp
@@ -47,12 +47,12 @@ class normal_herd : public linear_regression {
       storage_ptr storage);
   explicit normal_herd(storage_ptr storage);
 
-  void train(const common::sfv_t& fv, float value);
+  void train(const common::sfv_t& fv, double value);
   void update(
       const common::sfv_t& sfv,
-      float loss,
-      float variance,
-      float sign_error);
+      double loss,
+      double variance,
+      double sign_error);
 
   void clear();
